+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added 
 - `hueAsRad`, `hueAsGrad`, and `hueAsTurns` extensions for `HSL` and `HSV` classes.
 - `toHex()` now has a `renderAlpha` parameter that lets you render the color's alpha channel in the hex. By default the alpha will be added if it's < 1.  
+- `toCssRgb()` and `toCssHsl()` extensions to render colors as CSS color functions
 
 ### Changed
 - `withNumberSign` parameter of `toHex` now defaults to `true`